TST: GH26807 separate out test_keys, test_round_trip, test_append and test_file_handling
