+++ pandas/tests/io/pytables/test_store.py
@@ -1,5 +1,4 @@
 import datetime
-from datetime import timedelta
 from distutils.version import LooseVersion
 import hashlib
 from io import BytesIO
@@ -12,7 +11,7 @@ from warnings import catch_warnings, simplefilter
 import numpy as np
 import pytest
 
-from pandas.compat import is_platform_little_endian, is_platform_windows
+from pandas.compat import is_platform_windows
 import pandas.util._test_decorators as td
 
 import pandas as pd
@@ -43,7 +42,6 @@ from pandas.tests.io.pytables.common import (
 from pandas.io.pytables import (
     ClosedFileError,
     HDFStore,
-    PossibleDataLossError,
     Term,
     _maybe_adjust_name,
     read_hdf,
@@ -73,149 +71,6 @@ class TestHDFStore:
                 tbl["a"] = tm.makeDataFrame()
                 assert len(tbl) == 1
                 assert type(tbl["a"]) == DataFrame
-
-    def test_conv_read_write(self, setup_path):
-        with tm.ensure_clean() as path:
-
-            def roundtrip(key, obj, **kwargs):
-                obj.to_hdf(path, key, **kwargs)
-                return read_hdf(path, key)
-
-            o = tm.makeTimeSeries()
-            tm.assert_series_equal(o, roundtrip("series", o))
-
-            o = tm.makeStringSeries()
-            tm.assert_series_equal(o, roundtrip("string_series", o))
-
-            o = tm.makeDataFrame()
-            tm.assert_frame_equal(o, roundtrip("frame", o))
-
-            # table
-            df = DataFrame({"A": range(5), "B": range(5)})
-            df.to_hdf(path, "table", append=True)
-            result = read_hdf(path, "table", where=["index>2"])
-            tm.assert_frame_equal(df[df.index > 2], result)
-
-    def test_long_strings(self, setup_path):
-
-        # GH6166
-        df = DataFrame(
-            {"a": tm.rands_array(100, size=10)}, index=tm.rands_array(100, size=10)
-        )
-
-        with ensure_clean_store(setup_path) as store:
-            store.append("df", df, data_columns=["a"])
-
-            result = store.select("df")
-            tm.assert_frame_equal(df, result)
-
-    def test_api(self, setup_path):
-
-        # GH4584
-        # API issue when to_hdf doesn't accept append AND format args
-        with ensure_clean_path(setup_path) as path:
-
-            df = tm.makeDataFrame()
-            df.iloc[:10].to_hdf(path, "df", append=True, format="table")
-            df.iloc[10:].to_hdf(path, "df", append=True, format="table")
-            tm.assert_frame_equal(read_hdf(path, "df"), df)
-
-            # append to False
-            df.iloc[:10].to_hdf(path, "df", append=False, format="table")
-            df.iloc[10:].to_hdf(path, "df", append=True, format="table")
-            tm.assert_frame_equal(read_hdf(path, "df"), df)
-
-        with ensure_clean_path(setup_path) as path:
-
-            df = tm.makeDataFrame()
-            df.iloc[:10].to_hdf(path, "df", append=True)
-            df.iloc[10:].to_hdf(path, "df", append=True, format="table")
-            tm.assert_frame_equal(read_hdf(path, "df"), df)
-
-            # append to False
-            df.iloc[:10].to_hdf(path, "df", append=False, format="table")
-            df.iloc[10:].to_hdf(path, "df", append=True)
-            tm.assert_frame_equal(read_hdf(path, "df"), df)
-
-        with ensure_clean_path(setup_path) as path:
-
-            df = tm.makeDataFrame()
-            df.to_hdf(path, "df", append=False, format="fixed")
-            tm.assert_frame_equal(read_hdf(path, "df"), df)
-
-            df.to_hdf(path, "df", append=False, format="f")
-            tm.assert_frame_equal(read_hdf(path, "df"), df)
-
-            df.to_hdf(path, "df", append=False)
-            tm.assert_frame_equal(read_hdf(path, "df"), df)
-
-            df.to_hdf(path, "df")
-            tm.assert_frame_equal(read_hdf(path, "df"), df)
-
-        with ensure_clean_store(setup_path) as store:
-
-            df = tm.makeDataFrame()
-
-            _maybe_remove(store, "df")
-            store.append("df", df.iloc[:10], append=True, format="table")
-            store.append("df", df.iloc[10:], append=True, format="table")
-            tm.assert_frame_equal(store.select("df"), df)
-
-            # append to False
-            _maybe_remove(store, "df")
-            store.append("df", df.iloc[:10], append=False, format="table")
-            store.append("df", df.iloc[10:], append=True, format="table")
-            tm.assert_frame_equal(store.select("df"), df)
-
-            # formats
-            _maybe_remove(store, "df")
-            store.append("df", df.iloc[:10], append=False, format="table")
-            store.append("df", df.iloc[10:], append=True, format="table")
-            tm.assert_frame_equal(store.select("df"), df)
-
-            _maybe_remove(store, "df")
-            store.append("df", df.iloc[:10], append=False, format="table")
-            store.append("df", df.iloc[10:], append=True, format=None)
-            tm.assert_frame_equal(store.select("df"), df)
-
-        with ensure_clean_path(setup_path) as path:
-            # Invalid.
-            df = tm.makeDataFrame()
-
-            msg = "Can only append to Tables"
-
-            with pytest.raises(ValueError, match=msg):
-                df.to_hdf(path, "df", append=True, format="f")
-
-            with pytest.raises(ValueError, match=msg):
-                df.to_hdf(path, "df", append=True, format="fixed")
-
-            msg = r"invalid HDFStore format specified \[foo\]"
-
-            with pytest.raises(TypeError, match=msg):
-                df.to_hdf(path, "df", append=True, format="foo")
-
-            with pytest.raises(TypeError, match=msg):
-                df.to_hdf(path, "df", append=False, format="foo")
-
-        # File path doesn't exist
-        path = ""
-        msg = f"File {path} does not exist"
-
-        with pytest.raises(FileNotFoundError, match=msg):
-            read_hdf(path, "df")
-
-    def test_keys(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-            store["a"] = tm.makeTimeSeries()
-            store["b"] = tm.makeStringSeries()
-            store["c"] = tm.makeDataFrame()
-
-            assert len(store) == 3
-            expected = {"/a", "/b", "/c"}
-            assert set(store.keys()) == expected
-            assert set(store) == expected
 
     def test_no_track_times(self, setup_path):
 
@@ -259,57 +114,6 @@ class TestHDFStore:
 
         # checksums are NOT same if track_time = True
         assert checksum_0_tt_true != checksum_1_tt_true
-
-    def test_non_pandas_keys(self, setup_path):
-        class Table1(tables.IsDescription):
-            value1 = tables.Float32Col()
-
-        class Table2(tables.IsDescription):
-            value2 = tables.Float32Col()
-
-        class Table3(tables.IsDescription):
-            value3 = tables.Float32Col()
-
-        with ensure_clean_path(setup_path) as path:
-            with tables.open_file(path, mode="w") as h5file:
-                group = h5file.create_group("/", "group")
-                h5file.create_table(group, "table1", Table1, "Table 1")
-                h5file.create_table(group, "table2", Table2, "Table 2")
-                h5file.create_table(group, "table3", Table3, "Table 3")
-            with HDFStore(path) as store:
-                assert len(store.keys(include="native")) == 3
-                expected = {"/group/table1", "/group/table2", "/group/table3"}
-                assert set(store.keys(include="native")) == expected
-                assert set(store.keys(include="pandas")) == set()
-                for name in expected:
-                    df = store.get(name)
-                    assert len(df.columns) == 1
-
-    def test_keys_illegal_include_keyword_value(self, setup_path):
-        with ensure_clean_store(setup_path) as store:
-            with pytest.raises(
-                ValueError,
-                match="`include` should be either 'pandas' or 'native' "
-                "but is 'illegal'",
-            ):
-                store.keys(include="illegal")
-
-    def test_keys_ignore_hdf_softlink(self, setup_path):
-
-        # GH 20523
-        # Puts a softlink into HDF file and rereads
-
-        with ensure_clean_store(setup_path) as store:
-
-            df = DataFrame({"A": range(5), "B": range(5)})
-            store.put("df", df)
-
-            assert store.keys() == ["/df"]
-
-            store._handle.create_soft_link(store._handle.root, "symlink", "df")
-
-            # Should ignore the softlink
-            assert store.keys() == ["/df"]
 
     def test_iter_empty(self, setup_path):
 
@@ -408,167 +212,6 @@ class TestHDFStore:
             with pytest.raises(Exception, match=msg):
                 store.select("df2")
 
-    def test_mode(self, setup_path):
-
-        df = tm.makeTimeDataFrame()
-
-        def check(mode):
-
-            msg = r"[\S]* does not exist"
-            with ensure_clean_path(setup_path) as path:
-
-                # constructor
-                if mode in ["r", "r+"]:
-                    with pytest.raises(IOError, match=msg):
-                        HDFStore(path, mode=mode)
-
-                else:
-                    store = HDFStore(path, mode=mode)
-                    assert store._handle.mode == mode
-                    store.close()
-
-            with ensure_clean_path(setup_path) as path:
-
-                # context
-                if mode in ["r", "r+"]:
-                    with pytest.raises(IOError, match=msg):
-                        with HDFStore(path, mode=mode) as store:
-                            pass
-                else:
-                    with HDFStore(path, mode=mode) as store:
-                        assert store._handle.mode == mode
-
-            with ensure_clean_path(setup_path) as path:
-
-                # conv write
-                if mode in ["r", "r+"]:
-                    with pytest.raises(IOError, match=msg):
-                        df.to_hdf(path, "df", mode=mode)
-                    df.to_hdf(path, "df", mode="w")
-                else:
-                    df.to_hdf(path, "df", mode=mode)
-
-                # conv read
-                if mode in ["w"]:
-                    msg = (
-                        "mode w is not allowed while performing a read. "
-                        r"Allowed modes are r, r\+ and a."
-                    )
-                    with pytest.raises(ValueError, match=msg):
-                        read_hdf(path, "df", mode=mode)
-                else:
-                    result = read_hdf(path, "df", mode=mode)
-                    tm.assert_frame_equal(result, df)
-
-        def check_default_mode():
-
-            # read_hdf uses default mode
-            with ensure_clean_path(setup_path) as path:
-                df.to_hdf(path, "df", mode="w")
-                result = read_hdf(path, "df")
-                tm.assert_frame_equal(result, df)
-
-        check("r")
-        check("r+")
-        check("a")
-        check("w")
-        check_default_mode()
-
-    def test_reopen_handle(self, setup_path):
-
-        with ensure_clean_path(setup_path) as path:
-
-            store = HDFStore(path, mode="a")
-            store["a"] = tm.makeTimeSeries()
-
-            msg = (
-                r"Re-opening the file \[[\S]*\] with mode \[a\] will delete the "
-                "current file!"
-            )
-            # invalid mode change
-            with pytest.raises(PossibleDataLossError, match=msg):
-                store.open("w")
-
-            store.close()
-            assert not store.is_open
-
-            # truncation ok here
-            store.open("w")
-            assert store.is_open
-            assert len(store) == 0
-            store.close()
-            assert not store.is_open
-
-            store = HDFStore(path, mode="a")
-            store["a"] = tm.makeTimeSeries()
-
-            # reopen as read
-            store.open("r")
-            assert store.is_open
-            assert len(store) == 1
-            assert store._mode == "r"
-            store.close()
-            assert not store.is_open
-
-            # reopen as append
-            store.open("a")
-            assert store.is_open
-            assert len(store) == 1
-            assert store._mode == "a"
-            store.close()
-            assert not store.is_open
-
-            # reopen as append (again)
-            store.open("a")
-            assert store.is_open
-            assert len(store) == 1
-            assert store._mode == "a"
-            store.close()
-            assert not store.is_open
-
-    def test_open_args(self, setup_path):
-
-        with tm.ensure_clean(setup_path) as path:
-
-            df = tm.makeDataFrame()
-
-            # create an in memory store
-            store = HDFStore(
-                path, mode="a", driver="H5FD_CORE", driver_core_backing_store=0
-            )
-            store["df"] = df
-            store.append("df2", df)
-
-            tm.assert_frame_equal(store["df"], df)
-            tm.assert_frame_equal(store["df2"], df)
-
-            store.close()
-
-        # the file should not have actually been written
-        assert not os.path.exists(path)
-
-    def test_flush(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-            store["a"] = tm.makeTimeSeries()
-            store.flush()
-            store.flush(fsync=True)
-
-    def test_get(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-            store["a"] = tm.makeTimeSeries()
-            left = store.get("a")
-            right = store["a"]
-            tm.assert_series_equal(left, right)
-
-            left = store.get("/a")
-            right = store["/a"]
-            tm.assert_series_equal(left, right)
-
-            with pytest.raises(KeyError, match="'No object named b in the file'"):
-                store.get("b")
-
     @pytest.mark.parametrize(
         "where, expected",
         [
@@ -657,395 +300,6 @@ class TestHDFStore:
             for x in ["mode", "path", "handle", "complib"]:
                 getattr(store, f"_{x}")
 
-    def test_complibs_default_settings(self, setup_path):
-        # GH15943
-        df = tm.makeDataFrame()
-
-        # Set complevel and check if complib is automatically set to
-        # default value
-        with ensure_clean_path(setup_path) as tmpfile:
-            df.to_hdf(tmpfile, "df", complevel=9)
-            result = pd.read_hdf(tmpfile, "df")
-            tm.assert_frame_equal(result, df)
-
-            with tables.open_file(tmpfile, mode="r") as h5file:
-                for node in h5file.walk_nodes(where="/df", classname="Leaf"):
-                    assert node.filters.complevel == 9
-                    assert node.filters.complib == "zlib"
-
-        # Set complib and check to see if compression is disabled
-        with ensure_clean_path(setup_path) as tmpfile:
-            df.to_hdf(tmpfile, "df", complib="zlib")
-            result = pd.read_hdf(tmpfile, "df")
-            tm.assert_frame_equal(result, df)
-
-            with tables.open_file(tmpfile, mode="r") as h5file:
-                for node in h5file.walk_nodes(where="/df", classname="Leaf"):
-                    assert node.filters.complevel == 0
-                    assert node.filters.complib is None
-
-        # Check if not setting complib or complevel results in no compression
-        with ensure_clean_path(setup_path) as tmpfile:
-            df.to_hdf(tmpfile, "df")
-            result = pd.read_hdf(tmpfile, "df")
-            tm.assert_frame_equal(result, df)
-
-            with tables.open_file(tmpfile, mode="r") as h5file:
-                for node in h5file.walk_nodes(where="/df", classname="Leaf"):
-                    assert node.filters.complevel == 0
-                    assert node.filters.complib is None
-
-        # Check if file-defaults can be overridden on a per table basis
-        with ensure_clean_path(setup_path) as tmpfile:
-            store = HDFStore(tmpfile)
-            store.append("dfc", df, complevel=9, complib="blosc")
-            store.append("df", df)
-            store.close()
-
-            with tables.open_file(tmpfile, mode="r") as h5file:
-                for node in h5file.walk_nodes(where="/df", classname="Leaf"):
-                    assert node.filters.complevel == 0
-                    assert node.filters.complib is None
-                for node in h5file.walk_nodes(where="/dfc", classname="Leaf"):
-                    assert node.filters.complevel == 9
-                    assert node.filters.complib == "blosc"
-
-    def test_complibs(self, setup_path):
-        # GH14478
-        df = tm.makeDataFrame()
-
-        # Building list of all complibs and complevels tuples
-        all_complibs = tables.filters.all_complibs
-        # Remove lzo if its not available on this platform
-        if not tables.which_lib_version("lzo"):
-            all_complibs.remove("lzo")
-        # Remove bzip2 if its not available on this platform
-        if not tables.which_lib_version("bzip2"):
-            all_complibs.remove("bzip2")
-
-        all_levels = range(0, 10)
-        all_tests = [(lib, lvl) for lib in all_complibs for lvl in all_levels]
-
-        for (lib, lvl) in all_tests:
-            with ensure_clean_path(setup_path) as tmpfile:
-                gname = "foo"
-
-                # Write and read file to see if data is consistent
-                df.to_hdf(tmpfile, gname, complib=lib, complevel=lvl)
-                result = pd.read_hdf(tmpfile, gname)
-                tm.assert_frame_equal(result, df)
-
-                # Open file and check metadata
-                # for correct amount of compression
-                h5table = tables.open_file(tmpfile, mode="r")
-                for node in h5table.walk_nodes(where="/" + gname, classname="Leaf"):
-                    assert node.filters.complevel == lvl
-                    if lvl == 0:
-                        assert node.filters.complib is None
-                    else:
-                        assert node.filters.complib == lib
-                h5table.close()
-
-    def test_put_integer(self, setup_path):
-        # non-date, non-string index
-        df = DataFrame(np.random.randn(50, 100))
-        self._check_roundtrip(df, tm.assert_frame_equal, setup_path)
-
-    @pytest.mark.filterwarnings(
-        "ignore:object name:tables.exceptions.NaturalNameWarning"
-    )
-    def test_append(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-
-            # this is allowed by almost always don't want to do it
-            # tables.NaturalNameWarning):
-            with catch_warnings(record=True):
-
-                df = tm.makeTimeDataFrame()
-                _maybe_remove(store, "df1")
-                store.append("df1", df[:10])
-                store.append("df1", df[10:])
-                tm.assert_frame_equal(store["df1"], df)
-
-                _maybe_remove(store, "df2")
-                store.put("df2", df[:10], format="table")
-                store.append("df2", df[10:])
-                tm.assert_frame_equal(store["df2"], df)
-
-                _maybe_remove(store, "df3")
-                store.append("/df3", df[:10])
-                store.append("/df3", df[10:])
-                tm.assert_frame_equal(store["df3"], df)
-
-                # this is allowed by almost always don't want to do it
-                # tables.NaturalNameWarning
-                _maybe_remove(store, "/df3 foo")
-                store.append("/df3 foo", df[:10])
-                store.append("/df3 foo", df[10:])
-                tm.assert_frame_equal(store["df3 foo"], df)
-
-                # dtype issues - mizxed type in a single object column
-                df = DataFrame(data=[[1, 2], [0, 1], [1, 2], [0, 0]])
-                df["mixed_column"] = "testing"
-                df.loc[2, "mixed_column"] = np.nan
-                _maybe_remove(store, "df")
-                store.append("df", df)
-                tm.assert_frame_equal(store["df"], df)
-
-                # uints - test storage of uints
-                uint_data = DataFrame(
-                    {
-                        "u08": Series(
-                            np.random.randint(0, high=255, size=5), dtype=np.uint8
-                        ),
-                        "u16": Series(
-                            np.random.randint(0, high=65535, size=5), dtype=np.uint16
-                        ),
-                        "u32": Series(
-                            np.random.randint(0, high=2 ** 30, size=5), dtype=np.uint32
-                        ),
-                        "u64": Series(
-                            [2 ** 58, 2 ** 59, 2 ** 60, 2 ** 61, 2 ** 62],
-                            dtype=np.uint64,
-                        ),
-                    },
-                    index=np.arange(5),
-                )
-                _maybe_remove(store, "uints")
-                store.append("uints", uint_data)
-                tm.assert_frame_equal(store["uints"], uint_data)
-
-                # uints - test storage of uints in indexable columns
-                _maybe_remove(store, "uints")
-                # 64-bit indices not yet supported
-                store.append("uints", uint_data, data_columns=["u08", "u16", "u32"])
-                tm.assert_frame_equal(store["uints"], uint_data)
-
-    def test_append_series(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-
-            # basic
-            ss = tm.makeStringSeries()
-            ts = tm.makeTimeSeries()
-            ns = Series(np.arange(100))
-
-            store.append("ss", ss)
-            result = store["ss"]
-            tm.assert_series_equal(result, ss)
-            assert result.name is None
-
-            store.append("ts", ts)
-            result = store["ts"]
-            tm.assert_series_equal(result, ts)
-            assert result.name is None
-
-            ns.name = "foo"
-            store.append("ns", ns)
-            result = store["ns"]
-            tm.assert_series_equal(result, ns)
-            assert result.name == ns.name
-
-            # select on the values
-            expected = ns[ns > 60]
-            result = store.select("ns", "foo>60")
-            tm.assert_series_equal(result, expected)
-
-            # select on the index and values
-            expected = ns[(ns > 70) & (ns.index < 90)]
-            result = store.select("ns", "foo>70 and index<90")
-            tm.assert_series_equal(result, expected)
-
-            # multi-index
-            mi = DataFrame(np.random.randn(5, 1), columns=["A"])
-            mi["B"] = np.arange(len(mi))
-            mi["C"] = "foo"
-            mi.loc[3:5, "C"] = "bar"
-            mi.set_index(["C", "B"], inplace=True)
-            s = mi.stack()
-            s.index = s.index.droplevel(2)
-            store.append("mi", s)
-            tm.assert_series_equal(store["mi"], s)
-
-    @pytest.mark.skipif(
-        not is_platform_little_endian(), reason="reason platform is not little endian"
-    )
-    def test_encoding(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-            df = DataFrame({"A": "foo", "B": "bar"}, index=range(5))
-            df.loc[2, "A"] = np.nan
-            df.loc[3, "B"] = np.nan
-            _maybe_remove(store, "df")
-            store.append("df", df, encoding="ascii")
-            tm.assert_frame_equal(store["df"], df)
-
-            expected = df.reindex(columns=["A"])
-            result = store.select("df", Term("columns=A", encoding="ascii"))
-            tm.assert_frame_equal(result, expected)
-
-    @pytest.mark.parametrize(
-        "val",
-        [
-            [b"E\xc9, 17", b"", b"a", b"b", b"c"],
-            [b"E\xc9, 17", b"a", b"b", b"c"],
-            [b"EE, 17", b"", b"a", b"b", b"c"],
-            [b"E\xc9, 17", b"\xf8\xfc", b"a", b"b", b"c"],
-            [b"", b"a", b"b", b"c"],
-            [b"\xf8\xfc", b"a", b"b", b"c"],
-            [b"A\xf8\xfc", b"", b"a", b"b", b"c"],
-            [np.nan, b"", b"b", b"c"],
-            [b"A\xf8\xfc", np.nan, b"", b"b", b"c"],
-        ],
-    )
-    @pytest.mark.parametrize("dtype", ["category", object])
-    def test_latin_encoding(self, setup_path, dtype, val):
-        enc = "latin-1"
-        nan_rep = ""
-        key = "data"
-
-        val = [x.decode(enc) if isinstance(x, bytes) else x for x in val]
-        ser = Series(val, dtype=dtype)
-
-        with ensure_clean_path(setup_path) as store:
-            ser.to_hdf(store, key, format="table", encoding=enc, nan_rep=nan_rep)
-            retr = read_hdf(store, key)
-
-        s_nan = ser.replace(nan_rep, np.nan)
-
-        tm.assert_series_equal(s_nan, retr)
-
-    def test_append_some_nans(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-            df = DataFrame(
-                {
-                    "A": Series(np.random.randn(20)).astype("int32"),
-                    "A1": np.random.randn(20),
-                    "A2": np.random.randn(20),
-                    "B": "foo",
-                    "C": "bar",
-                    "D": Timestamp("20010101"),
-                    "E": datetime.datetime(2001, 1, 2, 0, 0),
-                },
-                index=np.arange(20),
-            )
-            # some nans
-            _maybe_remove(store, "df1")
-            df.loc[0:15, ["A1", "B", "D", "E"]] = np.nan
-            store.append("df1", df[:10])
-            store.append("df1", df[10:])
-            tm.assert_frame_equal(store["df1"], df)
-
-            # first column
-            df1 = df.copy()
-            df1.loc[:, "A1"] = np.nan
-            _maybe_remove(store, "df1")
-            store.append("df1", df1[:10])
-            store.append("df1", df1[10:])
-            tm.assert_frame_equal(store["df1"], df1)
-
-            # 2nd column
-            df2 = df.copy()
-            df2.loc[:, "A2"] = np.nan
-            _maybe_remove(store, "df2")
-            store.append("df2", df2[:10])
-            store.append("df2", df2[10:])
-            tm.assert_frame_equal(store["df2"], df2)
-
-            # datetimes
-            df3 = df.copy()
-            df3.loc[:, "E"] = np.nan
-            _maybe_remove(store, "df3")
-            store.append("df3", df3[:10])
-            store.append("df3", df3[10:])
-            tm.assert_frame_equal(store["df3"], df3)
-
-    def test_append_all_nans(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-
-            df = DataFrame(
-                {"A1": np.random.randn(20), "A2": np.random.randn(20)},
-                index=np.arange(20),
-            )
-            df.loc[0:15, :] = np.nan
-
-            # nan some entire rows (dropna=True)
-            _maybe_remove(store, "df")
-            store.append("df", df[:10], dropna=True)
-            store.append("df", df[10:], dropna=True)
-            tm.assert_frame_equal(store["df"], df[-4:])
-
-            # nan some entire rows (dropna=False)
-            _maybe_remove(store, "df2")
-            store.append("df2", df[:10], dropna=False)
-            store.append("df2", df[10:], dropna=False)
-            tm.assert_frame_equal(store["df2"], df)
-
-            # tests the option io.hdf.dropna_table
-            pd.set_option("io.hdf.dropna_table", False)
-            _maybe_remove(store, "df3")
-            store.append("df3", df[:10])
-            store.append("df3", df[10:])
-            tm.assert_frame_equal(store["df3"], df)
-
-            pd.set_option("io.hdf.dropna_table", True)
-            _maybe_remove(store, "df4")
-            store.append("df4", df[:10])
-            store.append("df4", df[10:])
-            tm.assert_frame_equal(store["df4"], df[-4:])
-
-            # nan some entire rows (string are still written!)
-            df = DataFrame(
-                {
-                    "A1": np.random.randn(20),
-                    "A2": np.random.randn(20),
-                    "B": "foo",
-                    "C": "bar",
-                },
-                index=np.arange(20),
-            )
-
-            df.loc[0:15, :] = np.nan
-
-            _maybe_remove(store, "df")
-            store.append("df", df[:10], dropna=True)
-            store.append("df", df[10:], dropna=True)
-            tm.assert_frame_equal(store["df"], df)
-
-            _maybe_remove(store, "df2")
-            store.append("df2", df[:10], dropna=False)
-            store.append("df2", df[10:], dropna=False)
-            tm.assert_frame_equal(store["df2"], df)
-
-            # nan some entire rows (but since we have dates they are still
-            # written!)
-            df = DataFrame(
-                {
-                    "A1": np.random.randn(20),
-                    "A2": np.random.randn(20),
-                    "B": "foo",
-                    "C": "bar",
-                    "D": Timestamp("20010101"),
-                    "E": datetime.datetime(2001, 1, 2, 0, 0),
-                },
-                index=np.arange(20),
-            )
-
-            df.loc[0:15, :] = np.nan
-
-            _maybe_remove(store, "df")
-            store.append("df", df[:10], dropna=True)
-            store.append("df", df[10:], dropna=True)
-            tm.assert_frame_equal(store["df"], df)
-
-            _maybe_remove(store, "df2")
-            store.append("df2", df[:10], dropna=False)
-            store.append("df2", df[10:], dropna=False)
-            tm.assert_frame_equal(store["df2"], df)
-
     def test_store_dropna(self, setup_path):
         df_with_missing = DataFrame(
             {"col1": [0.0, np.nan, 2.0], "col2": [1.0, np.nan, np.nan]},
@@ -1100,220 +354,6 @@ class TestHDFStore:
                 # still read from it.
                 pd.read_hdf(store, "k1")
 
-    def test_append_frame_column_oriented(self, setup_path):
-        with ensure_clean_store(setup_path) as store:
-
-            # column oriented
-            df = tm.makeTimeDataFrame()
-            df.index = df.index._with_freq(None)  # freq doesnt round-trip
-
-            _maybe_remove(store, "df1")
-            store.append("df1", df.iloc[:, :2], axes=["columns"])
-            store.append("df1", df.iloc[:, 2:])
-            tm.assert_frame_equal(store["df1"], df)
-
-            result = store.select("df1", "columns=A")
-            expected = df.reindex(columns=["A"])
-            tm.assert_frame_equal(expected, result)
-
-            # selection on the non-indexable
-            result = store.select("df1", ("columns=A", "index=df.index[0:4]"))
-            expected = df.reindex(columns=["A"], index=df.index[0:4])
-            tm.assert_frame_equal(expected, result)
-
-            # this isn't supported
-            msg = re.escape(
-                "passing a filterable condition to a non-table indexer "
-                "[Filter: Not Initialized]"
-            )
-            with pytest.raises(TypeError, match=msg):
-                store.select("df1", "columns=A and index>df.index[4]")
-
-    def test_append_with_different_block_ordering(self, setup_path):
-
-        # GH 4096; using same frames, but different block orderings
-        with ensure_clean_store(setup_path) as store:
-
-            for i in range(10):
-
-                df = DataFrame(np.random.randn(10, 2), columns=list("AB"))
-                df["index"] = range(10)
-                df["index"] += i * 10
-                df["int64"] = Series([1] * len(df), dtype="int64")
-                df["int16"] = Series([1] * len(df), dtype="int16")
-
-                if i % 2 == 0:
-                    del df["int64"]
-                    df["int64"] = Series([1] * len(df), dtype="int64")
-                if i % 3 == 0:
-                    a = df.pop("A")
-                    df["A"] = a
-
-                df.set_index("index", inplace=True)
-
-                store.append("df", df)
-
-        # test a different ordering but with more fields (like invalid
-        # combinate)
-        with ensure_clean_store(setup_path) as store:
-
-            df = DataFrame(np.random.randn(10, 2), columns=list("AB"), dtype="float64")
-            df["int64"] = Series([1] * len(df), dtype="int64")
-            df["int16"] = Series([1] * len(df), dtype="int16")
-            store.append("df", df)
-
-            # store additional fields in different blocks
-            df["int16_2"] = Series([1] * len(df), dtype="int16")
-            msg = re.escape(
-                "cannot match existing table structure for [int16] on appending data"
-            )
-            with pytest.raises(ValueError, match=msg):
-                store.append("df", df)
-
-            # store multiple additional fields in different blocks
-            df["float_3"] = Series([1.0] * len(df), dtype="float64")
-            msg = re.escape(
-                "cannot match existing table structure for [A,B] on appending data"
-            )
-            with pytest.raises(ValueError, match=msg):
-                store.append("df", df)
-
-    def test_append_with_strings(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-            with catch_warnings(record=True):
-
-                def check_col(key, name, size):
-                    assert (
-                        getattr(store.get_storer(key).table.description, name).itemsize
-                        == size
-                    )
-
-                # avoid truncation on elements
-                df = DataFrame([[123, "asdqwerty"], [345, "dggnhebbsdfbdfb"]])
-                store.append("df_big", df)
-                tm.assert_frame_equal(store.select("df_big"), df)
-                check_col("df_big", "values_block_1", 15)
-
-                # appending smaller string ok
-                df2 = DataFrame([[124, "asdqy"], [346, "dggnhefbdfb"]])
-                store.append("df_big", df2)
-                expected = concat([df, df2])
-                tm.assert_frame_equal(store.select("df_big"), expected)
-                check_col("df_big", "values_block_1", 15)
-
-                # avoid truncation on elements
-                df = DataFrame([[123, "asdqwerty"], [345, "dggnhebbsdfbdfb"]])
-                store.append("df_big2", df, min_itemsize={"values": 50})
-                tm.assert_frame_equal(store.select("df_big2"), df)
-                check_col("df_big2", "values_block_1", 50)
-
-                # bigger string on next append
-                store.append("df_new", df)
-                df_new = DataFrame(
-                    [[124, "abcdefqhij"], [346, "abcdefghijklmnopqrtsuvwxyz"]]
-                )
-                msg = (
-                    r"Trying to store a string with len \[26\] in "
-                    r"\[values_block_1\] column but\n"
-                    r"this column has a limit of \[15\]!\n"
-                    "Consider using min_itemsize to preset the sizes on these "
-                    "columns"
-                )
-                with pytest.raises(ValueError, match=msg):
-                    store.append("df_new", df_new)
-
-                # min_itemsize on Series index (GH 11412)
-                df = tm.makeMixedDataFrame().set_index("C")
-                store.append("ss", df["B"], min_itemsize={"index": 4})
-                tm.assert_series_equal(store.select("ss"), df["B"])
-
-                # same as above, with data_columns=True
-                store.append(
-                    "ss2", df["B"], data_columns=True, min_itemsize={"index": 4}
-                )
-                tm.assert_series_equal(store.select("ss2"), df["B"])
-
-                # min_itemsize in index without appending (GH 10381)
-                store.put("ss3", df, format="table", min_itemsize={"index": 6})
-                # just make sure there is a longer string:
-                df2 = df.copy().reset_index().assign(C="longer").set_index("C")
-                store.append("ss3", df2)
-                tm.assert_frame_equal(store.select("ss3"), pd.concat([df, df2]))
-
-                # same as above, with a Series
-                store.put("ss4", df["B"], format="table", min_itemsize={"index": 6})
-                store.append("ss4", df2["B"])
-                tm.assert_series_equal(
-                    store.select("ss4"), pd.concat([df["B"], df2["B"]])
-                )
-
-                # with nans
-                _maybe_remove(store, "df")
-                df = tm.makeTimeDataFrame()
-                df["string"] = "foo"
-                df.loc[df.index[1:4], "string"] = np.nan
-                df["string2"] = "bar"
-                df.loc[df.index[4:8], "string2"] = np.nan
-                df["string3"] = "bah"
-                df.loc[df.index[1:], "string3"] = np.nan
-                store.append("df", df)
-                result = store.select("df")
-                tm.assert_frame_equal(result, df)
-
-        with ensure_clean_store(setup_path) as store:
-
-            def check_col(key, name, size):
-                assert getattr(
-                    store.get_storer(key).table.description, name
-                ).itemsize, size
-
-            df = DataFrame({"A": "foo", "B": "bar"}, index=range(10))
-
-            # a min_itemsize that creates a data_column
-            _maybe_remove(store, "df")
-            store.append("df", df, min_itemsize={"A": 200})
-            check_col("df", "A", 200)
-            assert store.get_storer("df").data_columns == ["A"]
-
-            # a min_itemsize that creates a data_column2
-            _maybe_remove(store, "df")
-            store.append("df", df, data_columns=["B"], min_itemsize={"A": 200})
-            check_col("df", "A", 200)
-            assert store.get_storer("df").data_columns == ["B", "A"]
-
-            # a min_itemsize that creates a data_column2
-            _maybe_remove(store, "df")
-            store.append("df", df, data_columns=["B"], min_itemsize={"values": 200})
-            check_col("df", "B", 200)
-            check_col("df", "values_block_0", 200)
-            assert store.get_storer("df").data_columns == ["B"]
-
-            # infer the .typ on subsequent appends
-            _maybe_remove(store, "df")
-            store.append("df", df[:5], min_itemsize=200)
-            store.append("df", df[5:], min_itemsize=200)
-            tm.assert_frame_equal(store["df"], df)
-
-            # invalid min_itemsize keys
-            df = DataFrame(["foo", "foo", "foo", "barh", "barh", "barh"], columns=["A"])
-            _maybe_remove(store, "df")
-            msg = re.escape(
-                "min_itemsize has the key [foo] which is not an axis or data_column"
-            )
-            with pytest.raises(ValueError, match=msg):
-                store.append("df", df, min_itemsize={"foo": 20, "foobar": 20})
-
-    def test_append_with_empty_string(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-
-            # with all empty strings (GH 12242)
-            df = DataFrame({"x": ["a", "b", "c", "d", "e", "f", ""]})
-            store.append("df", df[:-1], min_itemsize={"x": 1})
-            store.append("df", df[-1:], min_itemsize={"x": 1})
-            tm.assert_frame_equal(store.select("df"), df)
-
     def test_to_hdf_with_min_itemsize(self, setup_path):
 
         with ensure_clean_path(setup_path) as path:
@@ -1344,163 +384,6 @@ class TestHDFStore:
 
             result = pd.read_hdf(path, "table", errors="surrogatepass")
             tm.assert_series_equal(result, ser)
-
-    def test_append_with_data_columns(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-            df = tm.makeTimeDataFrame()
-            df.iloc[0, df.columns.get_loc("B")] = 1.0
-            _maybe_remove(store, "df")
-            store.append("df", df[:2], data_columns=["B"])
-            store.append("df", df[2:])
-            tm.assert_frame_equal(store["df"], df)
-
-            # check that we have indices created
-            assert store._handle.root.df.table.cols.index.is_indexed is True
-            assert store._handle.root.df.table.cols.B.is_indexed is True
-
-            # data column searching
-            result = store.select("df", "B>0")
-            expected = df[df.B > 0]
-            tm.assert_frame_equal(result, expected)
-
-            # data column searching (with an indexable and a data_columns)
-            result = store.select("df", "B>0 and index>df.index[3]")
-            df_new = df.reindex(index=df.index[4:])
-            expected = df_new[df_new.B > 0]
-            tm.assert_frame_equal(result, expected)
-
-            # data column selection with a string data_column
-            df_new = df.copy()
-            df_new["string"] = "foo"
-            df_new.loc[df_new.index[1:4], "string"] = np.nan
-            df_new.loc[df_new.index[5:6], "string"] = "bar"
-            _maybe_remove(store, "df")
-            store.append("df", df_new, data_columns=["string"])
-            result = store.select("df", "string='foo'")
-            expected = df_new[df_new.string == "foo"]
-            tm.assert_frame_equal(result, expected)
-
-            # using min_itemsize and a data column
-            def check_col(key, name, size):
-                assert (
-                    getattr(store.get_storer(key).table.description, name).itemsize
-                    == size
-                )
-
-        with ensure_clean_store(setup_path) as store:
-            _maybe_remove(store, "df")
-            store.append(
-                "df", df_new, data_columns=["string"], min_itemsize={"string": 30}
-            )
-            check_col("df", "string", 30)
-            _maybe_remove(store, "df")
-            store.append("df", df_new, data_columns=["string"], min_itemsize=30)
-            check_col("df", "string", 30)
-            _maybe_remove(store, "df")
-            store.append(
-                "df", df_new, data_columns=["string"], min_itemsize={"values": 30}
-            )
-            check_col("df", "string", 30)
-
-        with ensure_clean_store(setup_path) as store:
-            df_new["string2"] = "foobarbah"
-            df_new["string_block1"] = "foobarbah1"
-            df_new["string_block2"] = "foobarbah2"
-            _maybe_remove(store, "df")
-            store.append(
-                "df",
-                df_new,
-                data_columns=["string", "string2"],
-                min_itemsize={"string": 30, "string2": 40, "values": 50},
-            )
-            check_col("df", "string", 30)
-            check_col("df", "string2", 40)
-            check_col("df", "values_block_1", 50)
-
-        with ensure_clean_store(setup_path) as store:
-            # multiple data columns
-            df_new = df.copy()
-            df_new.iloc[0, df_new.columns.get_loc("A")] = 1.0
-            df_new.iloc[0, df_new.columns.get_loc("B")] = -1.0
-            df_new["string"] = "foo"
-
-            sl = df_new.columns.get_loc("string")
-            df_new.iloc[1:4, sl] = np.nan
-            df_new.iloc[5:6, sl] = "bar"
-
-            df_new["string2"] = "foo"
-            sl = df_new.columns.get_loc("string2")
-            df_new.iloc[2:5, sl] = np.nan
-            df_new.iloc[7:8, sl] = "bar"
-            _maybe_remove(store, "df")
-            store.append("df", df_new, data_columns=["A", "B", "string", "string2"])
-            result = store.select(
-                "df", "string='foo' and string2='foo' and A>0 and B<0"
-            )
-            expected = df_new[
-                (df_new.string == "foo")
-                & (df_new.string2 == "foo")
-                & (df_new.A > 0)
-                & (df_new.B < 0)
-            ]
-            tm.assert_frame_equal(result, expected, check_freq=False)
-            # FIXME: 2020-05-07 freq check randomly fails in the CI
-
-            # yield an empty frame
-            result = store.select("df", "string='foo' and string2='cool'")
-            expected = df_new[(df_new.string == "foo") & (df_new.string2 == "cool")]
-            tm.assert_frame_equal(result, expected)
-
-        with ensure_clean_store(setup_path) as store:
-            # doc example
-            df_dc = df.copy()
-            df_dc["string"] = "foo"
-            df_dc.loc[df_dc.index[4:6], "string"] = np.nan
-            df_dc.loc[df_dc.index[7:9], "string"] = "bar"
-            df_dc["string2"] = "cool"
-            df_dc["datetime"] = Timestamp("20010102")
-            df_dc = df_dc._convert(datetime=True)
-            df_dc.loc[df_dc.index[3:5], ["A", "B", "datetime"]] = np.nan
-
-            _maybe_remove(store, "df_dc")
-            store.append(
-                "df_dc", df_dc, data_columns=["B", "C", "string", "string2", "datetime"]
-            )
-            result = store.select("df_dc", "B>0")
-
-            expected = df_dc[df_dc.B > 0]
-            tm.assert_frame_equal(result, expected)
-
-            result = store.select("df_dc", ["B > 0", "C > 0", "string == foo"])
-            expected = df_dc[(df_dc.B > 0) & (df_dc.C > 0) & (df_dc.string == "foo")]
-            tm.assert_frame_equal(result, expected, check_freq=False)
-            # FIXME: 2020-12-07 intermittent build failures here with freq of
-            #  None instead of BDay(4)
-
-        with ensure_clean_store(setup_path) as store:
-            # doc example part 2
-            np.random.seed(1234)
-            index = date_range("1/1/2000", periods=8)
-            df_dc = DataFrame(
-                np.random.randn(8, 3), index=index, columns=["A", "B", "C"]
-            )
-            df_dc["string"] = "foo"
-            df_dc.loc[df_dc.index[4:6], "string"] = np.nan
-            df_dc.loc[df_dc.index[7:9], "string"] = "bar"
-            df_dc.loc[:, ["B", "C"]] = df_dc.loc[:, ["B", "C"]].abs()
-            df_dc["string2"] = "cool"
-
-            # on-disk operations
-            store.append("df_dc", df_dc, data_columns=["B", "C", "string", "string2"])
-
-            result = store.select("df_dc", "B>0")
-            expected = df_dc[df_dc.B > 0]
-            tm.assert_frame_equal(result, expected)
-
-            result = store.select("df_dc", ["B > 0", "C > 0", 'string == "foo"'])
-            expected = df_dc[(df_dc.B > 0) & (df_dc.C > 0) & (df_dc.string == "foo")]
-            tm.assert_frame_equal(result, expected)
 
     def test_create_table_index(self, setup_path):
 
@@ -1566,30 +449,6 @@ class TestHDFStore:
                 with pytest.raises(AttributeError, match=msg):
                     store.create_table_index("f", columns=["string2"])
 
-    def test_append_hierarchical(self, setup_path):
-        index = MultiIndex(
-            levels=[["foo", "bar", "baz", "qux"], ["one", "two", "three"]],
-            codes=[[0, 0, 0, 1, 1, 2, 2, 3, 3, 3], [0, 1, 2, 0, 1, 1, 2, 0, 1, 2]],
-            names=["foo", "bar"],
-        )
-        df = DataFrame(np.random.randn(10, 3), index=index, columns=["A", "B", "C"])
-
-        with ensure_clean_store(setup_path) as store:
-            store.append("mi", df)
-            result = store.select("mi")
-            tm.assert_frame_equal(result, df)
-
-            # GH 3748
-            result = store.select("mi", columns=["A", "B"])
-            expected = df.reindex(columns=["A", "B"])
-            tm.assert_frame_equal(result, expected)
-
-        with ensure_clean_path("test.hdf") as path:
-            df.to_hdf(path, "df", format="table")
-            result = read_hdf(path, "df", columns=["A", "B"])
-            expected = df.reindex(columns=["A", "B"])
-            tm.assert_frame_equal(result, expected)
-
     def test_select_columns_in_where(self, setup_path):
 
         # GH 6169
@@ -1651,116 +510,6 @@ class TestHDFStore:
             with pytest.raises(TypeError, match=msg):
                 store.select("df", where=[("columns=A")])
 
-    def test_append_misc(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-            df = tm.makeDataFrame()
-            store.append("df", df, chunksize=1)
-            result = store.select("df")
-            tm.assert_frame_equal(result, df)
-
-            store.append("df1", df, expectedrows=10)
-            result = store.select("df1")
-            tm.assert_frame_equal(result, df)
-
-        # more chunksize in append tests
-        def check(obj, comparator):
-            for c in [10, 200, 1000]:
-                with ensure_clean_store(setup_path, mode="w") as store:
-                    store.append("obj", obj, chunksize=c)
-                    result = store.select("obj")
-                    comparator(result, obj)
-
-        df = tm.makeDataFrame()
-        df["string"] = "foo"
-        df["float322"] = 1.0
-        df["float322"] = df["float322"].astype("float32")
-        df["bool"] = df["float322"] > 0
-        df["time1"] = Timestamp("20130101")
-        df["time2"] = Timestamp("20130102")
-        check(df, tm.assert_frame_equal)
-
-        # empty frame, GH4273
-        with ensure_clean_store(setup_path) as store:
-
-            # 0 len
-            df_empty = DataFrame(columns=list("ABC"))
-            store.append("df", df_empty)
-            with pytest.raises(KeyError, match="'No object named df in the file'"):
-                store.select("df")
-
-            # repeated append of 0/non-zero frames
-            df = DataFrame(np.random.rand(10, 3), columns=list("ABC"))
-            store.append("df", df)
-            tm.assert_frame_equal(store.select("df"), df)
-            store.append("df", df_empty)
-            tm.assert_frame_equal(store.select("df"), df)
-
-            # store
-            df = DataFrame(columns=list("ABC"))
-            store.put("df2", df)
-            tm.assert_frame_equal(store.select("df2"), df)
-
-    def test_append_raise(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-
-            # test append with invalid input to get good error messages
-
-            # list in column
-            df = tm.makeDataFrame()
-            df["invalid"] = [["a"]] * len(df)
-            assert df.dtypes["invalid"] == np.object_
-            msg = re.escape(
-                """Cannot serialize the column [invalid]
-because its data contents are not [string] but [mixed] object dtype"""
-            )
-            with pytest.raises(TypeError, match=msg):
-                store.append("df", df)
-
-            # multiple invalid columns
-            df["invalid2"] = [["a"]] * len(df)
-            df["invalid3"] = [["a"]] * len(df)
-            with pytest.raises(TypeError, match=msg):
-                store.append("df", df)
-
-            # datetime with embedded nans as object
-            df = tm.makeDataFrame()
-            s = Series(datetime.datetime(2001, 1, 2), index=df.index)
-            s = s.astype(object)
-            s[0:5] = np.nan
-            df["invalid"] = s
-            assert df.dtypes["invalid"] == np.object_
-            msg = "too many timezones in this block, create separate data columns"
-            with pytest.raises(TypeError, match=msg):
-                store.append("df", df)
-
-            # directly ndarray
-            msg = "value must be None, Series, or DataFrame"
-            with pytest.raises(TypeError, match=msg):
-                store.append("df", np.arange(10))
-
-            # series directly
-            msg = re.escape(
-                "cannot properly create the storer for: "
-                "[group->df,value-><class 'pandas.core.series.Series'>]"
-            )
-            with pytest.raises(TypeError, match=msg):
-                store.append("df", Series(np.arange(10)))
-
-            # appending an incompatible table
-            df = tm.makeDataFrame()
-            store.append("df", df)
-
-            df["foo"] = "foo"
-            msg = re.escape(
-                "invalid combination of [non_index_axes] on appending data "
-                "[(1, ['A', 'B', 'C', 'D', 'foo'])] vs current table "
-                "[(1, ['A', 'B', 'C', 'D'])]"
-            )
-            with pytest.raises(ValueError, match=msg):
-                store.append("df", df)
-
     def test_table_index_incompatible_dtypes(self, setup_path):
         df1 = DataFrame({"a": [1, 2, 3]})
         df2 = DataFrame({"a": [4, 5, 6]}, index=date_range("1/1/2000", periods=3))
@@ -1770,70 +519,6 @@ because its data contents are not [string] but [mixed] object dtype"""
             msg = re.escape("incompatible kind in col [integer - datetime64]")
             with pytest.raises(TypeError, match=msg):
                 store.put("frame", df2, format="table", append=True)
-
-    def test_table_values_dtypes_roundtrip(self, setup_path):
-
-        with ensure_clean_store(setup_path) as store:
-            df1 = DataFrame({"a": [1, 2, 3]}, dtype="f8")
-            store.append("df_f8", df1)
-            tm.assert_series_equal(df1.dtypes, store["df_f8"].dtypes)
-
-            df2 = DataFrame({"a": [1, 2, 3]}, dtype="i8")
-            store.append("df_i8", df2)
-            tm.assert_series_equal(df2.dtypes, store["df_i8"].dtypes)
-
-            # incompatible dtype
-            msg = re.escape(
-                "invalid combination of [values_axes] on appending data "
-                "[name->values_block_0,cname->values_block_0,"
-                "dtype->float64,kind->float,shape->(1, 3)] vs "
-                "current table [name->values_block_0,"
-                "cname->values_block_0,dtype->int64,kind->integer,"
-                "shape->None]"
-            )
-            with pytest.raises(ValueError, match=msg):
-                store.append("df_i8", df1)
-
-            # check creation/storage/retrieval of float32 (a bit hacky to
-            # actually create them thought)
-            df1 = DataFrame(np.array([[1], [2], [3]], dtype="f4"), columns=["A"])
-            store.append("df_f4", df1)
-            tm.assert_series_equal(df1.dtypes, store["df_f4"].dtypes)
-            assert df1.dtypes[0] == "float32"
-
-            # check with mixed dtypes
-            df1 = DataFrame(
-                {
-                    c: Series(np.random.randint(5), dtype=c)
-                    for c in ["float32", "float64", "int32", "int64", "int16", "int8"]
-                }
-            )
-            df1["string"] = "foo"
-            df1["float322"] = 1.0
-            df1["float322"] = df1["float322"].astype("float32")
-            df1["bool"] = df1["float32"] > 0
-            df1["time1"] = Timestamp("20130101")
-            df1["time2"] = Timestamp("20130102")
-
-            store.append("df_mixed_dtypes1", df1)
-            result = store.select("df_mixed_dtypes1").dtypes.value_counts()
-            result.index = [str(i) for i in result.index]
-            expected = Series(
-                {
-                    "float32": 2,
-                    "float64": 1,
-                    "int32": 1,
-                    "bool": 1,
-                    "int16": 1,
-                    "int8": 1,
-                    "int64": 1,
-                    "object": 1,
-                    "datetime64[ns]": 2,
-                }
-            )
-            result = result.sort_index()
-            expected = expected.sort_index()
-            tm.assert_series_equal(result, expected)
 
     def test_table_mixed_dtypes(self, setup_path):
 
@@ -1914,55 +599,6 @@ because its data contents are not [string] but [date] object dtype"""
             store.append("table", s)
             result = store.select("table")
             tm.assert_series_equal(result, s)
-
-    def test_append_with_timedelta(self, setup_path):
-        # GH 3577
-        # append timedelta
-
-        df = DataFrame(
-            {
-                "A": Timestamp("20130101"),
-                "B": [
-                    Timestamp("20130101") + timedelta(days=i, seconds=10)
-                    for i in range(10)
-                ],
-            }
-        )
-        df["C"] = df["A"] - df["B"]
-        df.loc[3:5, "C"] = np.nan
-
-        with ensure_clean_store(setup_path) as store:
-
-            # table
-            _maybe_remove(store, "df")
-            store.append("df", df, data_columns=True)
-            result = store.select("df")
-            tm.assert_frame_equal(result, df)
-
-            result = store.select("df", where="C<100000")
-            tm.assert_frame_equal(result, df)
-
-            result = store.select("df", where="C<pd.Timedelta('-3D')")
-            tm.assert_frame_equal(result, df.iloc[3:])
-
-            result = store.select("df", "C<'-3D'")
-            tm.assert_frame_equal(result, df.iloc[3:])
-
-            # a bit hacky here as we don't really deal with the NaT properly
-
-            result = store.select("df", "C<'-500000s'")
-            result = result.dropna(subset=["C"])
-            tm.assert_frame_equal(result, df.iloc[6:])
-
-            result = store.select("df", "C<'-3.5D'")
-            result = result.iloc[1:]
-            tm.assert_frame_equal(result, df.iloc[4:])
-
-            # fixed
-            _maybe_remove(store, "df2")
-            store.put("df2", df)
-            result = store.select("df2")
-            tm.assert_frame_equal(result, df)
 
     def test_remove(self, setup_path):
 
@@ -2093,180 +729,6 @@ because its data contents are not [string] but [date] object dtype"""
             result = store.select("df", "index>datetime(2013,1,5)")
             tm.assert_frame_equal(result, expected)
 
-    def test_series(self, setup_path):
-
-        s = tm.makeStringSeries()
-        self._check_roundtrip(s, tm.assert_series_equal, path=setup_path)
-
-        ts = tm.makeTimeSeries()
-        self._check_roundtrip(ts, tm.assert_series_equal, path=setup_path)
-
-        ts2 = Series(ts.index, Index(ts.index, dtype=object))
-        self._check_roundtrip(ts2, tm.assert_series_equal, path=setup_path)
-
-        ts3 = Series(ts.values, Index(np.asarray(ts.index, dtype=object), dtype=object))
-        self._check_roundtrip(
-            ts3, tm.assert_series_equal, path=setup_path, check_index_type=False
-        )
-
-    def test_float_index(self, setup_path):
-
-        # GH #454
-        index = np.random.randn(10)
-        s = Series(np.random.randn(10), index=index)
-        self._check_roundtrip(s, tm.assert_series_equal, path=setup_path)
-
-    def test_tuple_index(self, setup_path):
-
-        # GH #492
-        col = np.arange(10)
-        idx = [(0.0, 1.0), (2.0, 3.0), (4.0, 5.0)]
-        data = np.random.randn(30).reshape((3, 10))
-        DF = DataFrame(data, index=idx, columns=col)
-
-        with catch_warnings(record=True):
-            simplefilter("ignore", pd.errors.PerformanceWarning)
-            self._check_roundtrip(DF, tm.assert_frame_equal, path=setup_path)
-
-    @pytest.mark.filterwarnings("ignore::pandas.errors.PerformanceWarning")
-    def test_index_types(self, setup_path):
-
-        with catch_warnings(record=True):
-            values = np.random.randn(2)
-
-            func = lambda l, r: tm.assert_series_equal(l, r, check_index_type=True)
-
-        with catch_warnings(record=True):
-            ser = Series(values, [0, "y"])
-            self._check_roundtrip(ser, func, path=setup_path)
-
-        with catch_warnings(record=True):
-            ser = Series(values, [datetime.datetime.today(), 0])
-            self._check_roundtrip(ser, func, path=setup_path)
-
-        with catch_warnings(record=True):
-            ser = Series(values, ["y", 0])
-            self._check_roundtrip(ser, func, path=setup_path)
-
-        with catch_warnings(record=True):
-            ser = Series(values, [datetime.date.today(), "a"])
-            self._check_roundtrip(ser, func, path=setup_path)
-
-        with catch_warnings(record=True):
-
-            ser = Series(values, [0, "y"])
-            self._check_roundtrip(ser, func, path=setup_path)
-
-            ser = Series(values, [datetime.datetime.today(), 0])
-            self._check_roundtrip(ser, func, path=setup_path)
-
-            ser = Series(values, ["y", 0])
-            self._check_roundtrip(ser, func, path=setup_path)
-
-            ser = Series(values, [datetime.date.today(), "a"])
-            self._check_roundtrip(ser, func, path=setup_path)
-
-            ser = Series(values, [1.23, "b"])
-            self._check_roundtrip(ser, func, path=setup_path)
-
-            ser = Series(values, [1, 1.53])
-            self._check_roundtrip(ser, func, path=setup_path)
-
-            ser = Series(values, [1, 5])
-            self._check_roundtrip(ser, func, path=setup_path)
-
-            ser = Series(
-                values, [datetime.datetime(2012, 1, 1), datetime.datetime(2012, 1, 2)]
-            )
-            self._check_roundtrip(ser, func, path=setup_path)
-
-    def test_timeseries_preepoch(self, setup_path):
-
-        dr = bdate_range("1/1/1940", "1/1/1960")
-        ts = Series(np.random.randn(len(dr)), index=dr)
-        try:
-            self._check_roundtrip(ts, tm.assert_series_equal, path=setup_path)
-        except OverflowError:
-            pytest.skip("known failer on some windows platforms")
-
-    @pytest.mark.parametrize(
-        "compression", [False, pytest.param(True, marks=td.skip_if_windows_python_3)]
-    )
-    def test_frame(self, compression, setup_path):
-
-        df = tm.makeDataFrame()
-
-        # put in some random NAs
-        df.values[0, 0] = np.nan
-        df.values[5, 3] = np.nan
-
-        self._check_roundtrip_table(
-            df, tm.assert_frame_equal, path=setup_path, compression=compression
-        )
-        self._check_roundtrip(
-            df, tm.assert_frame_equal, path=setup_path, compression=compression
-        )
-
-        tdf = tm.makeTimeDataFrame()
-        self._check_roundtrip(
-            tdf, tm.assert_frame_equal, path=setup_path, compression=compression
-        )
-
-        with ensure_clean_store(setup_path) as store:
-            # not consolidated
-            df["foo"] = np.random.randn(len(df))
-            store["df"] = df
-            recons = store["df"]
-            assert recons._mgr.is_consolidated()
-
-        # empty
-        self._check_roundtrip(df[:0], tm.assert_frame_equal, path=setup_path)
-
-    def test_empty_series_frame(self, setup_path):
-        s0 = Series(dtype=object)
-        s1 = Series(name="myseries", dtype=object)
-        df0 = DataFrame()
-        df1 = DataFrame(index=["a", "b", "c"])
-        df2 = DataFrame(columns=["d", "e", "f"])
-
-        self._check_roundtrip(s0, tm.assert_series_equal, path=setup_path)
-        self._check_roundtrip(s1, tm.assert_series_equal, path=setup_path)
-        self._check_roundtrip(df0, tm.assert_frame_equal, path=setup_path)
-        self._check_roundtrip(df1, tm.assert_frame_equal, path=setup_path)
-        self._check_roundtrip(df2, tm.assert_frame_equal, path=setup_path)
-
-    @pytest.mark.parametrize(
-        "dtype", [np.int64, np.float64, object, "m8[ns]", "M8[ns]"]
-    )
-    def test_empty_series(self, dtype, setup_path):
-        s = Series(dtype=dtype)
-        self._check_roundtrip(s, tm.assert_series_equal, path=setup_path)
-
-    def test_can_serialize_dates(self, setup_path):
-
-        rng = [x.date() for x in bdate_range("1/1/2000", "1/30/2000")]
-        frame = DataFrame(np.random.randn(len(rng), 4), index=rng)
-
-        self._check_roundtrip(frame, tm.assert_frame_equal, path=setup_path)
-
-    def test_store_hierarchical(self, setup_path):
-        index = MultiIndex(
-            levels=[["foo", "bar", "baz", "qux"], ["one", "two", "three"]],
-            codes=[[0, 0, 0, 1, 1, 2, 2, 3, 3, 3], [0, 1, 2, 0, 1, 1, 2, 0, 1, 2]],
-            names=["foo", "bar"],
-        )
-        frame = DataFrame(np.random.randn(10, 3), index=index, columns=["A", "B", "C"])
-
-        self._check_roundtrip(frame, tm.assert_frame_equal, path=setup_path)
-        self._check_roundtrip(frame.T, tm.assert_frame_equal, path=setup_path)
-        self._check_roundtrip(frame["A"], tm.assert_series_equal, path=setup_path)
-
-        # check that the names are stored
-        with ensure_clean_store(setup_path) as store:
-            store["frame"] = frame
-            recons = store["frame"]
-            tm.assert_frame_equal(recons, frame)
-
     def test_store_index_name(self, setup_path):
         df = tm.makeDataFrame()
         df.index.name = "foo"
@@ -2307,52 +769,6 @@ because its data contents are not [string] but [date] object dtype"""
             store["series"] = series
             recons = store["series"]
             tm.assert_series_equal(recons, series)
-
-    @pytest.mark.parametrize(
-        "compression", [False, pytest.param(True, marks=td.skip_if_windows_python_3)]
-    )
-    def test_store_mixed(self, compression, setup_path):
-        def _make_one():
-            df = tm.makeDataFrame()
-            df["obj1"] = "foo"
-            df["obj2"] = "bar"
-            df["bool1"] = df["A"] > 0
-            df["bool2"] = df["B"] > 0
-            df["int1"] = 1
-            df["int2"] = 2
-            return df._consolidate()
-
-        df1 = _make_one()
-        df2 = _make_one()
-
-        self._check_roundtrip(df1, tm.assert_frame_equal, path=setup_path)
-        self._check_roundtrip(df2, tm.assert_frame_equal, path=setup_path)
-
-        with ensure_clean_store(setup_path) as store:
-            store["obj"] = df1
-            tm.assert_frame_equal(store["obj"], df1)
-            store["obj"] = df2
-            tm.assert_frame_equal(store["obj"], df2)
-
-        # check that can store Series of all of these types
-        self._check_roundtrip(
-            df1["obj1"],
-            tm.assert_series_equal,
-            path=setup_path,
-            compression=compression,
-        )
-        self._check_roundtrip(
-            df1["bool1"],
-            tm.assert_series_equal,
-            path=setup_path,
-            compression=compression,
-        )
-        self._check_roundtrip(
-            df1["int1"],
-            tm.assert_series_equal,
-            path=setup_path,
-            compression=compression,
-        )
 
     @pytest.mark.filterwarnings(
         "ignore:\\nduplicate:pandas.io.pytables.DuplicateWarning"
@@ -3375,108 +1791,6 @@ because its data contents are not [string] but [date] object dtype"""
             expected = df[5:10]
             tm.assert_frame_equal(result, expected)
 
-    def test_append_to_multiple(self, setup_path):
-        df1 = tm.makeTimeDataFrame()
-        df2 = tm.makeTimeDataFrame().rename(columns="{}_2".format)
-        df2["foo"] = "bar"
-        df = concat([df1, df2], axis=1)
-
-        with ensure_clean_store(setup_path) as store:
-
-            # exceptions
-            msg = "append_to_multiple requires a selector that is in passed dict"
-            with pytest.raises(ValueError, match=msg):
-                store.append_to_multiple(
-                    {"df1": ["A", "B"], "df2": None}, df, selector="df3"
-                )
-
-            with pytest.raises(ValueError, match=msg):
-                store.append_to_multiple({"df1": None, "df2": None}, df, selector="df3")
-
-            msg = (
-                "append_to_multiple must have a dictionary specified as the way to "
-                "split the value"
-            )
-            with pytest.raises(ValueError, match=msg):
-                store.append_to_multiple("df1", df, "df1")
-
-            # regular operation
-            store.append_to_multiple(
-                {"df1": ["A", "B"], "df2": None}, df, selector="df1"
-            )
-            result = store.select_as_multiple(
-                ["df1", "df2"], where=["A>0", "B>0"], selector="df1"
-            )
-            expected = df[(df.A > 0) & (df.B > 0)]
-            tm.assert_frame_equal(result, expected)
-
-    def test_append_to_multiple_dropna(self, setup_path):
-        df1 = tm.makeTimeDataFrame()
-        df2 = tm.makeTimeDataFrame().rename(columns="{}_2".format)
-        df1.iloc[1, df1.columns.get_indexer(["A", "B"])] = np.nan
-        df = concat([df1, df2], axis=1)
-
-        with ensure_clean_store(setup_path) as store:
-
-            # dropna=True should guarantee rows are synchronized
-            store.append_to_multiple(
-                {"df1": ["A", "B"], "df2": None}, df, selector="df1", dropna=True
-            )
-            result = store.select_as_multiple(["df1", "df2"])
-            expected = df.dropna()
-            tm.assert_frame_equal(result, expected)
-            tm.assert_index_equal(store.select("df1").index, store.select("df2").index)
-
-    @pytest.mark.xfail(
-        run=False, reason="append_to_multiple_dropna_false is not raising as failed"
-    )
-    def test_append_to_multiple_dropna_false(self, setup_path):
-        df1 = tm.makeTimeDataFrame()
-        df2 = tm.makeTimeDataFrame().rename(columns="{}_2".format)
-        df1.iloc[1, df1.columns.get_indexer(["A", "B"])] = np.nan
-        df = concat([df1, df2], axis=1)
-
-        with ensure_clean_store(setup_path) as store:
-
-            # dropna=False shouldn't synchronize row indexes
-            store.append_to_multiple(
-                {"df1a": ["A", "B"], "df2a": None}, df, selector="df1a", dropna=False
-            )
-
-            # TODO Update error message to desired message for this case
-            msg = "Cannot select as multiple after appending with dropna=False"
-            with pytest.raises(ValueError, match=msg):
-                store.select_as_multiple(["df1a", "df2a"])
-
-            assert not store.select("df1a").index.equals(store.select("df2a").index)
-
-    def test_append_to_multiple_min_itemsize(self, setup_path):
-        # GH 11238
-        df = DataFrame(
-            {
-                "IX": np.arange(1, 21),
-                "Num": np.arange(1, 21),
-                "BigNum": np.arange(1, 21) * 88,
-                "Str": ["a" for _ in range(20)],
-                "LongStr": ["abcde" for _ in range(20)],
-            }
-        )
-        expected = df.iloc[[0]]
-
-        with ensure_clean_store(setup_path) as store:
-            store.append_to_multiple(
-                {
-                    "index": ["IX"],
-                    "nums": ["Num", "BigNum"],
-                    "strs": ["Str", "LongStr"],
-                },
-                df.iloc[[0]],
-                "index",
-                min_itemsize={"Str": 10, "LongStr": 100, "Num": 2},
-            )
-            result = store.select_as_multiple(["index", "nums", "strs"])
-            tm.assert_frame_equal(result, expected)
-
     def test_select_as_multiple(self, setup_path):
 
         df1 = tm.makeTimeDataFrame()
@@ -3721,43 +2035,6 @@ because its data contents are not [string] but [date] object dtype"""
 
         result = tm.round_trip_localpath(writer, reader)
         tm.assert_frame_equal(df, result)
-
-    def _check_roundtrip(self, obj, comparator, path, compression=False, **kwargs):
-
-        options = {}
-        if compression:
-            options["complib"] = _default_compressor
-
-        with ensure_clean_store(path, "w", **options) as store:
-            store["obj"] = obj
-            retrieved = store["obj"]
-            comparator(retrieved, obj, **kwargs)
-
-    def _check_double_roundtrip(
-        self, obj, comparator, path, compression=False, **kwargs
-    ):
-        options = {}
-        if compression:
-            options["complib"] = compression or _default_compressor
-
-        with ensure_clean_store(path, "w", **options) as store:
-            store["obj"] = obj
-            retrieved = store["obj"]
-            comparator(retrieved, obj, **kwargs)
-            store["obj"] = retrieved
-            again = store["obj"]
-            comparator(again, obj, **kwargs)
-
-    def _check_roundtrip_table(self, obj, comparator, path, compression=False):
-        options = {}
-        if compression:
-            options["complib"] = _default_compressor
-
-        with ensure_clean_store(path, "w", **options) as store:
-            store.put("obj", obj, format="table")
-            retrieved = store["obj"]
-
-            comparator(retrieved, obj)
 
     def test_multiple_open_close(self, setup_path):
         # gh-4409: open & close multiple times
@@ -4048,38 +2325,6 @@ because its data contents are not [string] but [date] object dtype"""
             assert result.index.freq == df.index.freq
             tm.assert_class_equal(result.index, df.index, obj="dataframe index")
 
-    def test_unicode_index(self, setup_path):
-
-        unicode_values = ["\u03c3", "\u03c3\u03c3"]
-
-        # PerformanceWarning
-        with catch_warnings(record=True):
-            simplefilter("ignore", pd.errors.PerformanceWarning)
-            s = Series(np.random.randn(len(unicode_values)), unicode_values)
-            self._check_roundtrip(s, tm.assert_series_equal, path=setup_path)
-
-    def test_unicode_longer_encoded(self, setup_path):
-        # GH 11234
-        char = "\u0394"
-        df = DataFrame({"A": [char]})
-        with ensure_clean_store(setup_path) as store:
-            store.put("df", df, format="table", encoding="utf-8")
-            result = store.get("df")
-            tm.assert_frame_equal(result, df)
-
-        df = DataFrame({"A": ["a", char], "B": ["b", "b"]})
-        with ensure_clean_store(setup_path) as store:
-            store.put("df", df, format="table", encoding="utf-8")
-            result = store.get("df")
-            tm.assert_frame_equal(result, df)
-
-    def test_store_datetime_mixed(self, setup_path):
-
-        df = DataFrame({"a": [1, 2, 3], "b": [1.0, 2.0, 3.0], "c": ["a", "b", "c"]})
-        ts = tm.makeTimeSeries()
-        df["d"] = ts.index[:3]
-        self._check_roundtrip(df, tm.assert_frame_equal, path=setup_path)
-
     # FIXME: don't leave commented-out code
     # def test_cant_write_multiindex_table(self):
     #     # for now, #1848
@@ -4306,17 +2551,6 @@ because its data contents are not [string] but [date] object dtype"""
             df.to_hdf(path, "df", format="table")
             other = read_hdf(path, "df")
 
-            tm.assert_frame_equal(df, other)
-            assert df.equals(other)
-            assert other.equals(df)
-
-    def test_round_trip_equals(self, setup_path):
-        # GH 9330
-        df = DataFrame({"B": [1, 2], "A": ["x", "y"]})
-
-        with ensure_clean_path(setup_path) as path:
-            df.to_hdf(path, "df", format="table")
-            other = read_hdf(path, "df")
             tm.assert_frame_equal(df, other)
             assert df.equals(other)
             assert other.equals(df)
